this is my first project
its just some python that tells you your sun sign if it has the day and month of your birthday. I'd like to make it more detailed with regards to the data it needs to out put that - like, degrees of a sign instead of it being based off days of the gregorian calendar but I'm still learning and I don't know if I'm ready to undertake making a database and creating variables that require a lot of information yet.